nvme: move key commands into new keys plugin
Consolidate gen-dhchap-key, check-dhchap-key, gen-tls-key,
check-tls-key, and tls-key into a "keys" plugin as gen-dhchap,
check-dhchap, gen-tls, check-tls, import, export, and revoke.

Splits tls-key's bundled --import/--export/--revoke modes into
separate subcommands, each with only the options it needs.

No backwards compatibility with the old command names is kept.
