+++ Documentation/cmds-main.txt
@@ -193,12 +193,6 @@ linknvme:nvme-boot-part-log[1]::
 linknvme:nvme-capacity-mgmt[1]::
 	Capacity Management Command
 
-linknvme:nvme-check-dhchap-key[1]::
-	Generate NVMeoF DH-HMAC-CHAP host key
-
-linknvme:nvme-check-tls-key[1]::
-	Validate NVMeoF TLS PSK
-
 linknvme:nvme-cmdset-ind-id-ns[1]::
 	I/O Command Set Independent Identify Namespace
 
@@ -208,14 +202,8 @@ linknvme:nvme-endurance-event-agg-log[1]::
 linknvme:nvme-fid-support-effects-log[1]::
 	Retrieve FID Support and Effects log
 
-linknvme:nvme-gen-dhchap-key[1]::
-	Generate NVMeoF DH-HMAC-CHAP host key
-
 linknvme:nvme-gen-hostnqn[1]::
 	Generate NVMeoF host NQN
-
-linknvme:nvme-gen-tls-key[1]::
-	Generate NVMeoF TLS PSK
 
 linknvme:nvme-get-lba-status[1]::
 	Get LBA Status command
