+++ Documentation/meson.build
@@ -77,6 +77,13 @@ adoc_sources = [
     'nvme-io-mgmt-recv',
     'nvme-io-mgmt-send',
     'nvme-io-passthru',
+    'nvme-keys-check-dhchap',
+    'nvme-keys-check-tls',
+    'nvme-keys-export',
+    'nvme-keys-gen-dhchap',
+    'nvme-keys-gen-tls',
+    'nvme-keys-import',
+    'nvme-keys-revoke',
     'nvme-lba-status-log',
     'nvme-list',
     'nvme-list-ctrl',
@@ -206,7 +213,6 @@ adoc_sources = [
     'nvme-subsystem-reset',
     'nvme-supported-log-pages',
     'nvme-telemetry-log',
-    'nvme-tls-key',
     'nvme-toshiba-clear-pcie-correctable-errors',
     'nvme-toshiba-vs-internal-log',
     'nvme-toshiba-vs-smart-add-log',
