--- Documentation/nvme-check-dhchap-key.1
@@ -1,51 +1,0 @@
-'\" t
-.\"     Title: nvme-check-dhchap-key
-.\"    Author: [FIXME: author] [see http://docbook.sf.net/el/author]
-.\" Generator: DocBook XSL Stylesheets v1.78.1 <http://docbook.sf.net/>
-.\"      Date: 11/25/2021
-.\"    Manual: NVMe Manual
-.\"    Source: NVMe
-.\"  Language: English
-.\"
-.TH "NVME\-CHECK\-DHCHAP\" "1" "11/25/2021" "NVMe" "NVMe Manual"
-.\" -----------------------------------------------------------------
-.\" * Define some portability stuff
-.\" -----------------------------------------------------------------
-.\" ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.\" http://bugs.debian.org/507673
-.\" http://lists.gnu.org/archive/html/groff/2009-02/msg00013.html
-.\" ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.ie \n(.g .ds Aq \(aq
-.el       .ds Aq '
-.\" -----------------------------------------------------------------
-.\" * set default formatting
-.\" -----------------------------------------------------------------
-.\" disable hyphenation
-.nh
-.\" disable justification (adjust text to left margin only)
-.ad l
-.\" -----------------------------------------------------------------
-.\" * MAIN CONTENT STARTS HERE *
-.\" -----------------------------------------------------------------
-.SH "NAME"
-nvme-check-dhchap-key \- Check a generated host DH\-HMAC\-CHAP key
-.SH "SYNOPSIS"
-.sp
-.nf
-\fInvme check\-dhchap\-key\fR [\-\-key=<key> ]
-.fi
-.SH "DESCRIPTION"
-.sp
-Checks if the key is a valid DH\-HMAC\-CHAP host key of the form: DHHC\-1:00:ia6zGodOr4SEG0Zzaw398rpY0wqipUWj4jWjUh4HWUz6aQ2n: and prints it to stdout\&.
-.SH "OPTIONS"
-.PP
-\-k <key>, \-\-key=<key>
-.RS 4
-Key to be checked\&.
-.RE
-.SH "EXAMPLES"
-.sp
-No Examples
-.SH "NVME"
-.sp
-Part of the nvme\-user suite
