--- Documentation/nvme-gen-dhchap-key.1
@@ -1,69 +1,0 @@
-'\" t
-.\"     Title: nvme-gen-dhchap-key
-.\"    Author: [FIXME: author] [see http://docbook.sf.net/el/author]
-.\" Generator: DocBook XSL Stylesheets v1.78.1 <http://docbook.sf.net/>
-.\"      Date: 11/25/2021
-.\"    Manual: NVMe Manual
-.\"    Source: NVMe
-.\"  Language: English
-.\"
-.TH "NVME\-GEN\-DHCHAP\-K" "1" "11/25/2021" "NVMe" "NVMe Manual"
-.\" -----------------------------------------------------------------
-.\" * Define some portability stuff
-.\" -----------------------------------------------------------------
-.\" ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.\" http://bugs.debian.org/507673
-.\" http://lists.gnu.org/archive/html/groff/2009-02/msg00013.html
-.\" ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.ie \n(.g .ds Aq \(aq
-.el       .ds Aq '
-.\" -----------------------------------------------------------------
-.\" * set default formatting
-.\" -----------------------------------------------------------------
-.\" disable hyphenation
-.nh
-.\" disable justification (adjust text to left margin only)
-.ad l
-.\" -----------------------------------------------------------------
-.\" * MAIN CONTENT STARTS HERE *
-.\" -----------------------------------------------------------------
-.SH "NAME"
-nvme-gen-dhchap-key \- Generate a host DH\-HMAC\-CHAP key
-.SH "SYNOPSIS"
-.sp
-.nf
-\fInvme gen\-dhchap\-key\fR [\-\-hmac=<hmac\-id> | \-h <hmac\-id>]
-                      [\-\-secret=<secret> | \-s <secret> ]
-                      [\-\-key\-length=<len> | \-l <len> ]
-                      [\-\-nqn=<host\-nqn> | \-n <host\-nqn> ]
-.fi
-.SH "DESCRIPTION"
-.sp
-Generate a base64\-encoded DH\-HMAC\-CHAP host key in the form: DHHC\-1:00:ia6zGodOr4SEG0Zzaw398rpY0wqipUWj4jWjUh4HWUz6aQ2n: and prints it to stdout\&.
-.SH "OPTIONS"
-.PP
-\-h <hmac\-id>, \-\-hmac=<hmac\-id>
-.RS 4
-Select a HMAC algorithm to use\&. Possible values are: 0 \- No HMAC algorithm 1 \- SHA\-256 2 \- SHA\-384 3 \- SHA\-512
-.RE
-.PP
-\-s <secret>, \-\-secret=<secret>
-.RS 4
-Secret value (in hexadecimal) to be used for the key\&. If none are provided a random value is used\&.
-.RE
-.PP
-\-l <len>, \-\-key\-length=<len>
-.RS 4
-Length of the resulting key\&. Possible values are 32, 48, or 64\&.
-.RE
-.PP
-\-n <hostnqn>, \-\-nqn=<hostnqn>
-.RS 4
-Host\-NQN to be used for the transformation\&. This parameter is only valid if a non\-zero HMAC function has been specified\&.
-.RE
-.SH "EXAMPLES"
-.sp
-No Examples
-.SH "NVME"
-.sp
-Part of the nvme\-user suite
