+++ Documentation/nvme-keys-export.txt
@@ -1,0 +1,64 @@
+nvme-keys-export(1)
+===================
+
+NAME
+----
+nvme-keys-export - Export NVMe TLS PSKs from a keyring
+
+SYNOPSIS
+--------
+[verse]
+'nvme' [<global-options>] 'keys export' [--keyring=<name> | -k <name>]
+			[--keyfile=<file> | -f <file>]
+
+DESCRIPTION
+-----------
+Export all NVMe TLS pre-shared keys (PSKs) from the system keyring
+in the form
+
+<description> <psk>
+
+where '<description>' is the key description from the exported key
+and '<psk>' is the key data in PSK interchange format
+'NVMeTLSkey-1:01:<base64 encoded data>:'. Each key is exported in a
+single line.
+
+OPTIONS
+-------
+-k <name>::
+--keyring=<name>::
+	Name of the keyring to export the retained keys from. Default
+	is '.nvme'.
+
+-f <file>::
+--keyfile=<file>::
+	File to write the exported keys to instead of stdout.
+
+include::global-options.txt[]
+
+EXAMPLES
+--------
+
+* Export previously created keys from the kernel keyring and store them into a file
++
+------------
+# nvme keys export -f nvme-tls-keys.txt
+------------
+
+* Export/list all keys from the .nvme keyring using nvme and keyctl
++
+------------
+# nvme keys export
+NVMe0R01 hostnqn0 subsys0 NVMeTLSkey-1:01:/b9tVz2OXJVISnoFgrPAygyS86XYJWkAapQeULns6PMpM8wv:
+
+# keyctl show
+Session Keyring
+ 573249525 --alswrv      0     0  keyring: _ses
+ 353599402 --alswrv      0 65534   \_ keyring: _uid.0
+ 475911922 ---lswrv      0     0   \_ keyring: .nvme
+ 649274894 --als-rv      0     0       \_ psk: NVMe0R01 hostnqn0 subsys0
+------------
+
+NVME
+----
+Part of the nvme-user suite
