+++ Documentation/nvme-keys-gen-tls.txt
@@ -1,18 +1,18 @@
-nvme-gen-tls-key(1)
-======================
+nvme-keys-gen-tls(1)
+====================
 
 NAME
 ----
-nvme-gen-tls-key - Generate a NVMe TLS PSK
+nvme-keys-gen-tls - Generate a NVMe TLS PSK
 
 SYNOPSIS
 --------
 [verse]
-'nvme' [<global-options>] 'gen-tls-key' [--keyring=<name> | -k <name>]
+'nvme' [<global-options>] 'keys gen-tls' [--keyring=<name> | -k <name>]
 			[--keytype=<type> | -t <type>]
 			[--hostnqn=<nqn> | -n <nqn>]
 			[--subsysnqn=<nqn> | -c <nqn>]
-			[--hmac=<hmac-id> | -h <hmac-id>]
+			[--hmac=<hmac-id> | -m <hmac-id>]
 			[--identity=<id-vers> | -I <id-vers>]
 			[--secret=<secret> | -s <secret>]
 			[--insert | -i]
@@ -61,7 +61,7 @@ OPTIONS
 	Subsystem NVMe Qualified Name (NQN) to be used to derive the
 	'retained' TLS key
 
--h <hmac-id>::
+-m <hmac-id>::
 --hmac=<hmac-id>::
 	Select a HMAC algorithm to use. Possible values are:
 	1 - SHA-256 (default)
@@ -83,15 +83,15 @@ OPTIONS
 	Insert the resulting TLS key into the keyring without printing out
 	the key in PSK interchange format.
 
--C:
---compat:
+-C::
+--compat::
 	Use the original non-RFC 8446 compliant algorithm when
-	deriving TLS keys for compatibility with older implentations.
+	deriving TLS keys for compatibility with older implementations.
 
--f <keyfile>
---keyfile=<keyfile>
-	Append the resulting TLS key to keyfile. This command line option is
-	depending on --insert.
+-f <keyfile>::
+--keyfile=<keyfile>::
+	Append the resulting TLS key to keyfile. This command line option
+	depends on --insert.
 
 include::global-options.txt[]
