+++ meson_options.txt
@@ -189,7 +189,7 @@ option(
   type : 'array',
   choices: ['amzn', 'config', 'dapustor', 'dell', 'dera',
             'exclusion', 'fdp', 'feat', 'huawei', 'ibm', 'innogrit', 'inspur',
-            'intel', 'lm', 'mangoboost', 'memblaze', 'micron', 'nbft',
+            'intel', 'keys', 'lm', 'mangoboost', 'memblaze', 'micron', 'nbft',
             'netapp', 'nvidia', 'ocp', 'registry', 'sandisk', 'scaleflux',
             'seagate', 'sed', 'shannon', 'solidigm', 'ssstc', 'toshiba',
             'transcend', 'utils', 'virtium', 'wdc', 'ymtc', 'zns'],
