+++ nvmf-autoconnect/udev-rules/70-nvmf-keys.rules.in
@@ -4,4 +4,4 @@
 #   the PSK keyring module gets loaded.
 #
 #
-ACTION=="add", SUBSYSTEM=="module", KERNEL=="nvme_tcp", TEST=="@SYSCONFDIR@/nvme/tls-keys", RUN+="@SBINDIR@/nvme tls --import --keyfile @SYSCONFDIR@/nvme/tls-keys"
+ACTION=="add", SUBSYSTEM=="module", KERNEL=="nvme_tcp", TEST=="@SYSCONFDIR@/nvme/tls-keys", RUN+="@SBINDIR@/nvme keys import --keyfile @SYSCONFDIR@/nvme/tls-keys"
